MatchTracker Pro (Pascal Mönch) 05.06.2023

+++ app/src/main/java/com/example/matchtrackerpro/data/datamodels/League.kt
@@ -2,7 +2,7 @@ package com.example.matchtrackerpro.data.datamodels
 
 
 data class League (
-    val leagueId: Long,
+    val leagueId: Int,
     val leagueName: String,
     val leagueImg: String,
     val teams: List<Team>
